+++ tests/snowpark/test_function.py
@@ -52,6 +52,7 @@ def test_deploy_function(
         dedent(
             """\
             create or replace function MockDatabase.MockSchema.func1(a string default 'default value', b variant)
+            copy grants
             returns string
             language python
             runtime_version=3.10
@@ -99,6 +100,7 @@ def test_deploy_function_with_external_access(
         dedent(
             """\
             create or replace function MockDatabase.MockSchema.func1(a string, b variant)
+            copy grants
             returns string
             language python
             runtime_version=3.8
@@ -218,6 +220,7 @@ def test_deploy_function_needs_update_because_packages_changes(
         dedent(
             """\
             create or replace function MockDatabase.MockSchema.func1(a string default 'default value', b variant)
+            copy grants
             returns string
             language python
             runtime_version=3.10
@@ -268,6 +271,7 @@ def test_deploy_function_needs_update_because_handler_changes(
         dedent(
             """\
             create or replace function MockDatabase.MockSchema.func1(a string default 'default value', b variant)
+            copy grants
             returns string
             language python
             runtime_version=3.10